+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 新增
+
+- 发送面板输入框中 `ESC` 键关闭
+- 发送面板输入框中增加确认快捷键
+  - macOs：`Cmd/Ctrl + Enter`
+  - Other：`Ctrl + Enter`
+
 ## 优化
 
 - 发送面板弹出后输入框自动获取焦点